Not sure, it's been a while :)
git-svn-id: http://svn.rockethands.com/rocketsauce/kranzky/Projects/dBoard@512 619292ac-9744-4b63-ad4b-f05b7053f485

+++ node.hpp
@@ -19,6 +19,6 @@ public:
 private:
     char m_symbol;
     unsigned char m_num;
-    unsigned short m_count;
+    unsigned int m_count;
     Node ** m_child;
 };
